+++ src/main/java/freemind/view/mindmapview/RootMainView.java
@@ -55,14 +55,14 @@ class RootMainView extends MainView {
     public void paintDragOver(Graphics2D graphics) {
         final int draggedOver = getDraggedOver();
         if (draggedOver == NodeView.DRAGGED_OVER_SON) {
-            graphics.setPaint(new GradientPaint(getWidth() / 4, 0,
-                    getNodeView().getMap().getBackground(), getWidth() * 3 / 4,
+            graphics.setPaint(new GradientPaint((float) getWidth() / 4, 0,
+                    getNodeView().getMap().getBackground(), (float) (getWidth() * 3) / 4,
                     0, NodeView.dragColor));
             graphics.fillRect(getWidth() / 4, 0, getWidth() - 1,
                     getHeight() - 1);
         } else if (draggedOver == NodeView.DRAGGED_OVER_SON_LEFT) {
-            graphics.setPaint(new GradientPaint(getWidth() * 3 / 4, 0,
-                    getNodeView().getMap().getBackground(), getWidth() / 4, 0,
+            graphics.setPaint(new GradientPaint((float) (getWidth() * 3) / 4, 0,
+                    getNodeView().getMap().getBackground(), (float) getWidth() / 4, 0,
                     NodeView.dragColor));
             graphics.fillRect(0, 0, getWidth() * 3 / 4, getHeight() - 1);
         }
@@ -141,7 +141,7 @@ class RootMainView extends MainView {
      * @return true if should be on the left, false otherwise.
      */
     public boolean dropPosition(double xCoord) {
-        return xCoord < getSize().width * 1 / 2;
+        return xCoord < (double) (getSize().width * 1) / 2;
     }
 
 }
